+++ src/main/java/com/excilys/cdb/configuration/MainConfig.java
@@ -2,7 +2,6 @@ package com.excilys.cdb.configuration;
 
 import javax.sql.DataSource;
 
-import org.apache.log4j.BasicConfigurator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -10,16 +9,22 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.web.context.AbstractContextLoaderInitializer;
-import org.springframework.web.context.WebApplicationContext;
+import org.springframework.web.WebApplicationInitializer;
 import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
+import org.springframework.web.servlet.DispatcherServlet;
+import org.springframework.web.servlet.ViewResolver;
+import org.springframework.web.servlet.view.InternalResourceViewResolver;
+import org.springframework.web.servlet.view.JstlView;
 import org.springframework.jdbc.datasource.DriverManagerDataSource;
 
+import javax.servlet.ServletContext;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRegistration;
 
 @Configuration
-@ComponentScan("com.excilys.cdb")
+@ComponentScan("com.excilys.cdb") /// ajouter les package a scanner
 @PropertySource("classpath:datasourcemysql.properties")
-public class MainConfig  extends AbstractContextLoaderInitializer {
+public class MainConfig implements WebApplicationInitializer {
 	
 	@Value("${dataSource.driverClassName}")
 	private String jdbcDriver;
@@ -33,12 +38,15 @@ public class MainConfig  extends AbstractContextLoaderInitializer {
 	private static final Logger LOG = LoggerFactory.getLogger(MainConfig.class );	
 	
 	@Override
-	protected WebApplicationContext createRootApplicationContext() {
-		AnnotationConfigWebApplicationContext rootContext = new AnnotationConfigWebApplicationContext();
-		rootContext.register(MainConfig.class);
-		LOG.info("Spring starting ...");
-		return rootContext;
-	}
+	public void onStartup(ServletContext ctx) throws ServletException {
+        AnnotationConfigWebApplicationContext webCtx = new AnnotationConfigWebApplicationContext();
+        webCtx.register(MainConfig.class);
+        webCtx.setServletContext(ctx);
+        
+        ServletRegistration.Dynamic servlet = ctx.addServlet("dispatcher", new DispatcherServlet(webCtx));
+        servlet.setLoadOnStartup(1);
+        servlet.addMapping("/");
+    }
 	
 	@Bean
     public DataSource dataSource() {
@@ -47,8 +55,18 @@ public class MainConfig  extends AbstractContextLoaderInitializer {
         dataSource.setUrl(url);
         dataSource.setUsername(username);
         dataSource.setPassword(password);
- 
         return dataSource;
     }
 	
+	@Bean
+  public ViewResolver viewResolver() {
+      InternalResourceViewResolver viewResolver = new InternalResourceViewResolver();
+      viewResolver.setViewClass(JstlView.class);
+      viewResolver.setPrefix("/WEB-INF/views/");
+      viewResolver.setSuffix(".jsp");
+      return viewResolver;
+  }
+	
 }
+
+
