+++ src/main/java/com/excilys/cdb/servlet/AddComputerServlet.java
@@ -3,6 +3,7 @@ package com.excilys.cdb.servlet;
 import java.io.IOException;
 import java.util.List;
 
+import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -11,7 +12,8 @@ import javax.servlet.http.HttpServletResponse;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.context.support.SpringBeanAutowiringSupport;
 
 import com.excilys.cdb.model.Company;
 import com.excilys.cdb.model.Computer;
@@ -24,10 +26,19 @@ import com.excilys.cdb.validator.Validator;
 public class AddComputerServlet extends HttpServlet{
 	
 	private static final long serialVersionUID = 1L;
-	private ComputerService computerService ;
+	@Autowired
+	private ComputerService computerService;
+	@Autowired
 	private CompanyService companyService ;
 	private List<Company> listCompany;
 	private static final Logger LOGGER = LoggerFactory.getLogger(AddComputerServlet.class);	
+	
+
+	@Override
+	public void init(ServletConfig config) throws ServletException {
+		super.init(config);
+		SpringBeanAutowiringSupport.processInjectionBasedOnCurrentContext(this);
+	}
 	
 	@Override
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
