+++ src/main/webapp/WEB-INF/web.xml
@@ -3,7 +3,28 @@
 <web-app>
 
 
-	<servlet>
+ 	<!-- <display-name>Spring MVC Application</display-name>
+ 	
+ 	<servlet>
+	  <servlet-name>springDispatcherServlet</servlet-name>
+	  <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
+	 
+	</servlet> -->
+<!--
+   <servlet>
+      <servlet-name>HelloWeb</servlet-name>
+      <servlet-class>
+         org.springframework.web.servlet.DispatcherServlet
+      </servlet-class>
+      <load-on-startup>1</load-on-startup>
+   </servlet>
+
+   <servlet-mapping>
+      <servlet-name>HelloWeb</servlet-name>
+      <url-pattern>/</url-pattern>
+   </servlet-mapping> -->
+
+<!-- 	<servlet>
 		<servlet-name>ListComputersServlet</servlet-name>
 		<servlet-class>com.excilys.cdb.servlet.ListComputersServlet</servlet-class>
 	</servlet>
@@ -20,6 +41,6 @@
 	<servlet-mapping>
 		<servlet-name>AddComputerServlet</servlet-name>
 		<url-pattern>/addcomputer</url-pattern>
-	</servlet-mapping>
+	</servlet-mapping> -->
 	
 </web-app>
